Created a service for users

+++ src/schemas/habit.ts
@@ -1,7 +1,7 @@
 import * as z from "zod"
 
 export const HabitSchema = z.object({
-    id: z.coerce.number(),
+    id: z.coerce.number().int(),
     name: z.string().max(30),
     frequency: z.enum(["DAILY", "WEEKLY", "N_PER_WEEK"]),
     frequencyPerWeek: z.number().int().positive().min(1).max(7).nullable(),
